Fix Firefox issues with OGDI. Fixes #6
Error is due to the bing maps API whose call an auto loaded script
(AtlasCompat.js) that is not properly instanced with Firefox.
A second error occured after applying this fix because of the
MicrosoftAjax.js script. The results is an infinite refresh loop on the
page. To solve this issue the MicrosoftAjax.js script is replaced by the
new script coming from ASP.NET MVC 4.
The detailed fix description can be found here :
http://stackoverflow.com/questions/7110023/firefox-6-infinite-page-refresh-with-page-with-hash-tags